requirements.txt
Django==1.11
mysqlclient==1.3.12
Pillow==5.0.0
django-jet==1.0.7
django-ckeditor==5.4.0
sorl-thumbnail==12.4.1
